+++ src/domains/Parser/classes/Parser/model/Parser.ts
@@ -1,7 +1,18 @@
+import { runTimerRegex, TimerRegexResult } from "../../TimerRegex";
+import fromRegexResultToOutput from "../lib/fromRegexResultToOutput";
+import regexes from "../lib/regexes";
 import ParserInput from "./ParserInput";
 import ParserOptions, { defaultParserOptions } from "./ParserOptions";
 import ParserOutput from "./ParserOutput";
-import _parse from "../lib/parse";
+
+function doRegex(str: string): TimerRegexResult | undefined {
+  for (const regex of regexes) {
+    const res = runTimerRegex(regex, str);
+    if (res) return res;
+  }
+
+  return undefined;
+}
 
 export default class Parser {
   private opt: ParserOptions;
@@ -11,6 +22,20 @@ export default class Parser {
   }
 
   parse(input: ParserInput): ParserOutput {
-    return _parse(input, this.opt);
+    const { slugs } = input;
+    if (slugs.length > 1) {
+      throw new Error("Unimplemented");
+    }
+
+    const regexed = doRegex(slugs[0]);
+    if (regexed === undefined) {
+      return {
+        type: "fail",
+      };
+    }
+
+    const output = fromRegexResultToOutput(regexed);
+
+    return output;
   }
 }
