+++ src/domains/Parser/classes/TimerRegex/index.ts
@@ -1,0 +1,2 @@
+export * from "./model";
+export { default as runTimerRegex } from "./lib/runRegexMatcher";
