+++ src/domains/Parser/classes/TimerRegex/model/index.ts
@@ -1,0 +1,10 @@
+export interface TimerRegexData {
+  pattern: RegExp;
+  groupKeys: (keyof TimerRegexResult)[];
+}
+export interface TimerRegexResult {
+  hours?: string;
+  minutes?: string;
+  seconds?: string;
+  times?: string;
+}
